Move Functionality in Accumulo
