+++ accumulo/src/main/scala/geotrellis/spark/io/accumulo/AccumuloLayerReader.scala
@@ -33,6 +33,8 @@ class AccumuloLayerReader(val attributeStore: AttributeStore)(implicit sc: Spark
 
   val defaultNumPartitions = sc.defaultParallelism
 
+  def sparkContext: SparkContext = sc
+
   def read[
     K: AvroRecordCodec: Boundable: JsonFormat: ClassTag,
     V: AvroRecordCodec: ClassTag,
